Pull submit buttons into own file
